% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlst-package.R
\docType{package}
\name{mnlst-package}
\alias{mnlst-package}
\alias{rstanlm}
\title{The 'mnlst' package.}
\description{
Write and Fit Moderated Nonlinear Latent State Trait Models with R and Stan.

A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munchfab/mnlst}
  \item Report bugs at \url{https://github.com/munchfab/mnlst/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/munchfab/mnlst}
  \item Report bugs at \url{https://github.com/munchfab/mnlst/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
